requirements_caption.txt
# Qwen 2.5 VL Caption Service Requirements (GGUF)
llama-cpp-python>=0.2.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

